requirements.txt
aiohttp==3.8.0
aiosignal==1.2.0
altgraph==0.17.2
astroid==2.8.4
async-timeout==4.0.0
attrs==21.2.0
autopep8==1.6.0
backoff==1.11.1
bandit==1.7.0
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.0.7
colorama==0.4.4
decorator==5.1.0
et-xmlfile==1.1.0
flake8==4.0.1
frozenlist==1.2.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.24
idna==3.3
importlib-metadata==4.8.1
isort==5.10.0
jdcal==1.4.1
lazy-object-proxy==1.6.0
lxml==4.6.4
macholib==1.15.2
mccabe==0.6.1
multidict==5.1.0
numpy==1.21.3
openpyxl==3.0.9
pandas==1.3.4
pbr==5.6.0
pefile==2021.9.3
Pillow==8.4.0
platformdirs==2.4.0
pycodestyle==2.8.0
pycountry==20.7.3
pyflakes==2.4.0
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pylint==2.11.1
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
six==1.16.0
smmap==5.0.0
stevedore==3.5.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.62.3
typed-ast==1.5.1
typing-extensions==3.10.0.2
urllib3==1.26.7
validators==0.18.2
wrapt==1.13.3
wxasync==0.43
wxPython==4.1.1
yarl==1.7.2
zipp==3.6.0
